+++ Taxify.Web/Controllers/HomeController.cs
@@ -1,89 +1,32 @@
-﻿using Microsoft.AspNetCore.Authentication.Cookies;
-using Microsoft.AspNetCore.Authentication;
-using Microsoft.AspNetCore.Authorization;
-using Microsoft.AspNetCore.Mvc;
-using System.Diagnostics;
-using System.Security.Claims;
-using Taxify.Service.DTOs.Users;
-using Taxify.Service.Interfaces;
+﻿using System.Diagnostics;
 using Taxify.Web.Models;
+using Microsoft.AspNetCore.Mvc;
+using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Authentication;
+using Microsoft.AspNetCore.Authentication.Cookies;
 
 namespace Taxify.Web.Controllers;
 
 public class HomeController : Controller
 {
-    private readonly IUserService userService;
-    private readonly ILogger<HomeController> _logger;
+    private readonly ILogger<HomeController> logger;
 
-    public HomeController(ILogger<HomeController> logger, IUserService userService)
+    public HomeController(ILogger<HomeController> logger)
     {
-        _logger = logger;
-        this.userService = userService;
+        this.logger = logger;
     }
 
-    public IActionResult Index(LoginModel model)
-    {
-        ClaimsPrincipal claimUser = HttpContext.User;
-
-        if (claimUser.Identity.IsAuthenticated)
-            return RedirectToAction("Main", "Home");
-
-        return View(model);
-    }
-
-    [HttpPost]
-    public async Task<IActionResult> Login(LoginModel model)
-    {
-        var userLoginDto = new UserLoginDto
-        {
-            Password = model.Password,
-            Phone = model.Phone
-        };
-
-        try
-        {
-            var user = await userService.LoginAsync(userLoginDto);
-            if (user is not null)
-            { 
-                List<Claim> claims = new List<Claim>() { 
-                    new Claim(ClaimTypes.OtherPhone, model.Phone),
-                    new Claim("OtherProperties","Example Role")
-                
-                };
-
-                ClaimsIdentity claimsIdentity = new ClaimsIdentity(claims,
-                    CookieAuthenticationDefaults.AuthenticationScheme );
-
-                AuthenticationProperties properties = new AuthenticationProperties() { 
-                
-                    AllowRefresh = true,
-                    IsPersistent = model.KeepLoggedIn
-                };
-
-                await HttpContext.SignInAsync(CookieAuthenticationDefaults.AuthenticationScheme,
-                    new ClaimsPrincipal(claimsIdentity), properties);
-            
-                return RedirectToAction("Main", "Home");
-
-            }
-        }
-        catch(Exception ex)
-        {
-            TempData["Message"] = ex.Message;
-        }
-        return RedirectToAction(actionName:"Index", routeValues: model);
-    }
     [Authorize]
     public IActionResult Main()
     {
         return View();
     }
 
-     public async Task<IActionResult> LogOut()
-     {
-         await HttpContext.SignOutAsync(CookieAuthenticationDefaults.AuthenticationScheme);
-         return RedirectToAction("Index","Home");
-     }
+    public async Task<IActionResult> Logout()
+    {
+        await HttpContext.SignOutAsync(CookieAuthenticationDefaults.AuthenticationScheme);
+        return RedirectToAction("Index", "Auth");
+    }
 
 
     public IActionResult Privacy()
