+++ Taxify.Web/Views/Auth/Index.cshtml
