+++ Taxify.Web/Views/Home/Main.cshtml
@@ -1,15 +1,114 @@
-﻿@*
-    For more information on enabling MVC for empty projects, visit https://go.microsoft.com/fwlink/?LinkID=397860
-*@
-@{
+﻿@{
+    Layout = "Menubar";
 }
 
+    <div class="collapse h-100 w-100" id="navbarToggleExternalContent">
+        <div class="position-absolute start-0 col-12 col-md-4 mx-auto p-3 shadow h-100 w-25 bg-dark" style="color:white">
+            <div class="menu accordion" style="color:white">
+                <ul class="menu-list">
+                    <li>
+                        <a class="menu-item" href="#" style="color:white"><i class="fa-solid fa-house me-2"></i> Home</a>
+                    </li>
 
+                    <li>
+                        <div class="menu-item">
+                            <button class="accordion-button collapsed p-0 fw-normal bg-transparent text-reset" type="button" data-bs-toggle="collapse" aria-expanded="false" data-bs-target="#submenu-1">
+                                <i class="fa-solid fa-chart-column me-2"></i> Analytics
+                            </button>
+                        </div>
+                        <div class="accordion-collapse collapse" id="submenu-1">
+                            <ul class="menu-list">
+                                <li>
+                                    <a class="menu-item" href="#">Insights</a>
+                                </li>
+                                <li>
+                                    <a class="menu-item" href="#">Reports</a>
+                                </li>
+                            </ul>
+                        </div>
+                    </li>
+                    <li>
+                        <div class="menu-item">
+                            <button class="accordion-button p-0 fw-normal bg-transparent text-reset" type="button" data-bs-toggle="collapse" aria-expanded="false" data-bs-target="#submenu-2">
+                                <i class="fa-solid fa-tag me-2"></i> Drivers
+                            </button>
+                        </div>
+                        <div class="accordion-collapse collapse" id="submenu-2">
+                            <ul class="menu-list">
+                                <li>
+                                    <a class="menu-item" href="#">#dashbord</a>
+                                </li>
+                                <li>
+                                    <a class="menu-item" href="#">#dashbord</a>
+                                </li>
+                                <li>
+                                    <a class="menu-item" href="#">#dashbord</a>
+                                </li>
+                            </ul>
+                        </div>
+                    </li>
+                    <li>
+                        <div class="menu-item">
+                            <button class="accordion-button p-0 fw-normal bg-transparent text-reset" type="button" data-bs-toggle="collapse" aria-expanded="false" data-bs-target="#submenu-3">
+                                <i class="fa-solid fa-tag me-2"></i> Users
+                            </button>
+                        </div>
+                        <div class="accordion-collapse collapse" id="submenu-3">
+                            <ul class="menu-list">
+                                <li>
+                                    <a class="menu-item" href="#">#dashbord</a>
+                                </li>
+                                <li>
+                                    <a class="menu-item" href="#">#dashbord</a>
+                                </li>
+                                <li>
+                                    <a class="menu-item" href="#">#dashbord</a>
+                                </li>
+                            </ul>
+                        </div>
+                    </li>
 
-<h1>
-    Hellow admin panel
-</h1>
-
-<form asp-action="Logout">
-    <button class="btn-dark" type="submit">Log out</button>
-</form>
+                    <li>
+                        <div class="menu-item">
+                            <button class="accordion-button p-0 fw-normal bg-transparent text-reset" type="button" data-bs-toggle="collapse" aria-expanded="false" data-bs-target="#submenu-5">
+                                <i class="fa-solid fa-tag me-2"></i> Orders
+                            </button>
+                        </div>
+                        <div class="accordion-collapse collapse" id="submenu-5">
+                            <ul class="menu-list">
+                                <li>
+                                    <a class="menu-item" href="#">#dashbord</a>
+                                </li>
+                                <li>
+                                    <a class="menu-item" href="#">#dashbord</a>
+                                </li>
+                                <li>
+                                    <a class="menu-item" href="#">#dashbord</a>
+                                </li>
+                            </ul>
+                        </div>
+                    </li>
+                    <li>
+                        <div class="menu-item">
+                            <button class="accordion-button p-0 fw-normal bg-transparent text-reset" type="button" data-bs-toggle="collapse" aria-expanded="false" data-bs-target="#submenu-6">
+                                <i class="fa-solid fa-tag me-2"></i> Others
+                            </button>
+                        </div>
+                        <div class="accordion-collapse collapse" id="submenu-6">
+                            <ul class="menu-list">
+                                <li>
+                                    <a class="menu-item" href="#">#dashbord</a>
+                                </li>
+                                <li>
+                                    <a class="menu-item" href="#">#dashbord</a>
+                                </li>
+                                <li>
+                                    <a class="menu-item" href="#">#dashbord</a>
+                                </li>
+                            </ul>
+                        </div>
+                    </li>
+                </ul>
+            </div>
+        </div>
+    </div>
