+++ Taxify.Web/Views/Shared/libaries.cshtml
@@ -4,7 +4,7 @@
     <meta charset="utf-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>@ViewData["Title"] - Taxify.Web</title>
-    <link rel="stylesheet" href="~/lib/MDBootstrap/css/mdb.min.css"/>
+    <link rel="stylesheet" href="~/lib/MDBootstrap/css/mdb.min.css" />
     <link rel="stylesheet" href="~/css/site.css" asp-append-version="true" />
     <link rel="stylesheet" href="~/Taxify.Web.styles.css" asp-append-version="true" />
 </head>
